Generate code from RBFS 22.2.1

+++ pkg/rbfs/state/model_bgp_instance.go
@@ -13,7 +13,7 @@ package state
 type BgpInstance struct {
 	InstanceName string `json:"instance_name,omitempty"`
 	// The AS number.
-	Asn int32 `json:"asn,omitempty"`
+	Asn int64 `json:"asn,omitempty"`
 	// Unique router ID
 	RouterId string `json:"router_id,omitempty"`
 	// The BGP cluster ID this router is member of.
